+++ mod-source-record-manager-server/src/main/java/org/folio/dao/JobMonitoringDaoImpl.java
@@ -1,5 +1,19 @@
 package org.folio.dao;
 
+import static java.lang.String.format;
+
+import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
+
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
 import io.vertx.sqlclient.Row;
@@ -7,28 +21,21 @@ import io.vertx.sqlclient.RowSet;
 import io.vertx.sqlclient.Tuple;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.folio.dao.util.PostgresClientFactory;
-import org.folio.rest.jaxrs.model.JobMonitoring;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.Date;
-import java.util.Iterator;
-import java.util.Optional;
-import java.util.UUID;
-
-import static java.lang.String.format;
-import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
+import org.folio.dao.util.PostgresClientFactory;
+import org.folio.rest.jaxrs.model.JobMonitoring;
 
 @Repository
 public class JobMonitoringDaoImpl implements JobMonitoringDao {
+
   private static final Logger LOGGER = LogManager.getLogger();
   private static final String TABLE = "job_monitoring";
   private static final String INSERT_SQL = "INSERT INTO %s.%s (id, job_execution_id, last_event_timestamp, notification_sent) VALUES ($1, $2, $3, $4)";
   private static final String SELECT_BY_JOB_EXECUTION_ID_QUERY = "SELECT id, job_execution_id, last_event_timestamp, notification_sent FROM %s.%s WHERE job_execution_id = $1";
+  private static final String SELECT_ALL_JOB_MONITORING_QUERY = "SELECT id, job_execution_id, last_event_timestamp, notification_sent FROM %s.%s";
+  private static final String SELECT_BY_NOTIFICATION_BEFORE_TIMESTAMP_QUERY = "SELECT * FROM %s.%s WHERE last_event_timestamp <= $1 AND notification_sent = $2";
   private static final String UPDATE_BY_JOB_EXECUTION_ID_QUERY = "UPDATE %s.%s SET last_event_timestamp = $1, notification_sent = $2 WHERE job_execution_id = $3";
   private static final String DELETE_BY_JOB_EXECUTION_ID_QUERY = "DELETE FROM %s.%s WHERE job_execution_id = $1";
 
@@ -44,18 +51,45 @@ public class JobMonitoringDaoImpl implements JobMonitoringDao {
     return promise.future().map(this::mapRowToJobMonitoring);
   }
 
+  @Override
+  public Future<List<JobMonitoring>> findAll(String tenantId) {
+    Promise<RowSet<Row>> promise = Promise.promise();
+    String query = format(SELECT_ALL_JOB_MONITORING_QUERY, convertToPsqlStandard(tenantId), TABLE);
+    pgClientFactory.createInstance(tenantId).select(query, promise);
+    return promise.future().map(this::mapResultSetToJobMonitoringList);
+  }
+
+  @Override
+  public Future<List<JobMonitoring>> findByNotificationBeforeTimestamp(LocalDateTime timestamp, boolean notificationSent,
+                                                                       String tenantId) {
+    Promise<RowSet<Row>> promise = Promise.promise();
+    String query = format(SELECT_BY_NOTIFICATION_BEFORE_TIMESTAMP_QUERY, convertToPsqlStandard(tenantId), TABLE);
+    Tuple queryParams = Tuple.of(timestamp, notificationSent);
+    pgClientFactory.createInstance(tenantId).select(query, queryParams, promise);
+    return promise.future().map(this::mapResultSetToJobMonitoringList);
+  }
+
+  private List<JobMonitoring> mapResultSetToJobMonitoringList(RowSet<Row> resultSet) {
+    List<JobMonitoring> jobMonitors = new ArrayList<>();
+    resultSet.forEach(row -> jobMonitors.add(mapRowToJobMonitoring(row)));
+    return jobMonitors;
+  }
+
+  private JobMonitoring mapRowToJobMonitoring(Row row) {
+    return new JobMonitoring()
+      .withId(row.getValue("id").toString())
+      .withJobExecutionId(row.getValue("job_execution_id").toString())
+      .withLastEventTimestamp(Date.from(row.getLocalDateTime("last_event_timestamp").toInstant(ZoneOffset.UTC)))
+      .withNotificationSent(row.getBoolean("notification_sent"));
+  }
+
   private Optional<JobMonitoring> mapRowToJobMonitoring(RowSet<Row> resultSet) {
     Iterator<Row> iterator = resultSet.iterator();
     if (!iterator.hasNext()) {
       return Optional.empty();
     } else {
       Row row = iterator.next();
-      JobMonitoring jobMonitoring = new JobMonitoring();
-      jobMonitoring.setId(row.getUUID("id").toString());
-      jobMonitoring.setJobExecutionId(row.getUUID("job_execution_id").toString());
-      jobMonitoring.setLastEventTimestamp(Date.from(row.getLocalDateTime("last_event_timestamp").toInstant(ZoneOffset.UTC)));
-      jobMonitoring.setNotificationSent(row.getBoolean("notification_sent"));
-      return Optional.of(jobMonitoring);
+      return Optional.of(mapRowToJobMonitoring(row));
     }
   }