+++ mod-source-record-manager-server/src/main/java/org/folio/rest/impl/ModTenantAPI.java
@@ -1,31 +1,29 @@
 package org.folio.rest.impl;
 
-import io.vertx.core.AsyncResult;
+import java.util.Map;
+
 import io.vertx.core.Context;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.Promise;
 import io.vertx.core.Vertx;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.RowSet;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.Map;
-import javax.ws.rs.core.Response;
-
-import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.TenantAttributes;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.services.MappingRuleService;
 import org.folio.spring.SpringContextUtil;
 import org.folio.util.pubsub.PubSubClientUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 
 public class ModTenantAPI extends TenantAPI {
+
   private static final Logger LOGGER = LogManager.getLogger();
 
-  private static final String GRANT_SEQUENCES_PERMISSION_PATTERN = "GRANT USAGE, SELECT ON ALL SEQUENCES IN SCHEMA %s TO %s;";
+  private static final String GRANT_SEQUENCES_PERMISSION_PATTERN =
+    "GRANT USAGE, SELECT ON ALL SEQUENCES IN SCHEMA %s TO %s;";
 
   @Autowired
   private MappingRuleService mappingRuleService;
@@ -40,7 +38,15 @@ public class ModTenantAPI extends TenantAPI {
     return super.loadData(attributes, tenantId, headers, context)
       .compose(num -> setSequencesPermissionForDbUser(context, tenantId)
         .compose(ar -> mappingRuleService.saveDefaultRules(tenantId))
-        .compose(ar -> registerModuleToPubsub(headers, context.owner())).map(num));
+        .compose(ar -> registerModuleToPubsub(headers, context.owner()))
+        .compose(ar -> saveTenantId(tenantId, context))
+        .map(num));
+  }
+
+  private Future<Void> saveTenantId(String tenantId, Context context) {
+    Vertx owner = context.owner();
+    owner.sharedData().getLocalMap("tenants").put(tenantId, tenantId);
+    return Future.succeededFuture();
   }
 
   private Future<RowSet<Row>> setSequencesPermissionForDbUser(Context context, String tenantId) {